Use separated component for pill copy
