+++ web-app/packages/admin-lib/src/common/components/index.ts
@@ -2,5 +2,4 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
 
-export * from '@/common'
-export * from '@/modules'
+export { default as CopyPill } from './CopyPill.vue'
