+++ web-app/packages/lib/src/common/composables/index.ts
@@ -3,6 +3,5 @@
 // SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-MerginMaps-Commercial
 
 export { default as useRouterTitle } from './use_router_title'
-export { default as useCopyToClipboard } from './useCopyToClipboard'
 export { useDataTableSearch } from './useDataTableSearch'
 export type { DataTableSearchOptions } from './useDataTableSearch'
